chore: move AddModule extension method to BuildingBlocks
