+++ BuildingBlocks/src/BuildingBlocks.API/Extensions/ServiceCollectionExtensions.cs
@@ -1,5 +1,7 @@
 ﻿using Autofac;
 using Autofac.Extensions.DependencyInjection;
+using Microsoft.AspNetCore.Mvc.ApplicationParts;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Enmeshed.BuildingBlocks.API.Extensions;
@@ -33,5 +35,23 @@ public static class ServiceCollectionExtensions
             default:
                 throw new Exception($"Unsupported database provider: {provider}");
         }
+    }
+
+    public static IServiceCollection AddModule<TModule>(this IServiceCollection services, IConfiguration configuration)
+        where TModule : IModule, new()
+    {
+        // Register assembly in MVC so it can find controllers of the module
+        services.AddControllers().ConfigureApplicationPartManager(manager =>
+            manager.ApplicationParts.Add(new AssemblyPart(typeof(TModule).Assembly)));
+
+        var module = new TModule();
+
+        var moduleConfiguration = configuration.GetSection($"Modules:{module.Name}");
+
+        module.ConfigureServices(services, moduleConfiguration);
+
+        services.AddSingleton<IModule>(module);
+
+        return services;
     }
 }
